+++ devp2p/kademlia.py
@@ -23,7 +23,7 @@ from devp2p import slogging
 import logging
 from .crypto import sha3
 from .utils import big_endian_to_int
-from quarkchain.tools.rlp_utils import encode_hex, is_integer, str_to_bytes
+from quarkchain.rlp.utils import encode_hex, is_integer, str_to_bytes
 
 log = slogging.get_logger('p2p.discovery.kademlia')
