+++ devp2p/multiplexer.py
@@ -1,7 +1,7 @@
 from gevent.queue import Queue
 from collections import OrderedDict
 import rlp
-from quarkchain.tools.rlp_utils import str_to_bytes, is_integer
+from quarkchain.rlp.utils import str_to_bytes, is_integer
 import struct
 import sys
