+++ devp2p/tests/test_discovery.py
@@ -2,7 +2,7 @@ from devp2p import discovery
 from devp2p import kademlia
 from devp2p import crypto
 from devp2p.app import BaseApp
-from quarkchain.tools.rlp_utils import decode_hex, encode_hex
+from quarkchain.rlp.utils import decode_hex, encode_hex
 from devp2p.utils import remove_chars
 import pytest
 import gevent
