+++ devp2p/tests/test_ecies.py
@@ -1,5 +1,5 @@
 import devp2p.crypto as crypto
-from quarkchain.tools.rlp_utils import decode_hex
+from quarkchain.rlp.utils import decode_hex
 
 
 def test_ecies_enc():
