+++ devp2p/tests/test_go_handshake.py
@@ -4,7 +4,7 @@
 
 from devp2p.rlpxcipher import RLPxSession
 from devp2p.crypto import ECCx, privtopub
-from quarkchain.tools.rlp_utils import decode_hex
+from quarkchain.rlp.utils import decode_hex
 test_values = \
     {
         "initiator_private_key": "5e173f6ac3c669587538e7727cf19b782a4f2fda07c1eaa662c593e5e85e3051",
