+++ devp2p/tests/test_go_signature.py
@@ -1,5 +1,5 @@
 from devp2p.crypto import ecdsa_sign, mk_privkey, privtopub, ecdsa_recover, ECCx
-from quarkchain.tools.rlp_utils import decode_hex
+from quarkchain.rlp.utils import decode_hex
 import pyelliptic
